Translator/requirements.txt
streamlit               
langchain       
langchain-google-genai    
python-dotenv              
langchain-core     
google-auth 